Dodanie walidacji, edycja widoku nawigacji dla admina, optymalizacja i porzadkowanie kodu

+++ src/Controller/CategoryController.php
@@ -11,44 +11,27 @@ use App\Repository\CategoryRepository;
 use http\Env\Response;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
- * Class CategoryController
- *
+ * Class CategoryController.
  */
 class CategoryController extends AbstractController
 {
-
     /**
      * Index action.
      *
      * @param Request            $request
      * @param CategoryRepository $categoryRepository
      * @param PaginatorInterface $paginator
-     * @param int                $pageNumber
      *
      * @return Response|\Symfony\Component\HttpFoundation\Response
      *
-     * @Route("/categories/{pageNumber}", name="categories", defaults={"pageNumber"=1}, methods={"GET"})
+     * @Route("/categories", name="categories", methods={"GET"})
      */
-    public function index(Request $request, CategoryRepository $categoryRepository, PaginatorInterface $paginator, int $pageNumber)
+    public function index(Request $request, CategoryRepository $categoryRepository, PaginatorInterface $paginator)
     {
-//        $limit = 3;
-//        $categories = $this->getDoctrine()->getRepository(Category::class)->findBy(
-//            [],
-//            ['id' => 'desc'],
-//            $limit,
-//            ($pageNumber - 1) * $limit
-//        );
-
-//        return $this->render('category/index.html.twig', [
-//            'categories' => $categories,
-//            'pageNumber' => $pageNumber,
-//        ]);
-
         $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
         $pagination = $paginator->paginate(
@@ -72,7 +55,7 @@ class CategoryController extends AbstractController
      *
      * @throws \Exception
      *
-     * @Route("/category/create", name="category_create")
+     * @Route("/category/create", name="category_create", methods={"GET", "POST"})
      */
     public function create(Request $request)
     {
@@ -80,7 +63,6 @@ class CategoryController extends AbstractController
 
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
-        $form->add('save', SubmitType::class);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $category = $form->getData();
@@ -99,24 +81,26 @@ class CategoryController extends AbstractController
         );
     }
 
-
     /**
      * Update action.
      *
-     * @param \Symfony\Component\HttpFoundation\Request $request
-     * @param int                                       $id
-     * @param Category                                  $category
-     * @param CategoryRepository                        $repository
+     * @param Request            $request
+     * @param Category           $category
+     * @param CategoryRepository $repository
+     * @param int                $id
      *
      * @return Response|\Symfony\Component\HttpFoundation\Response
      *
-     * @Route("/category/{id}/update", name="category_update", methods={"GET","PUT"})
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
+     *
+     * @Route("/category/{id}/update", name="category_update", requirements={"id": "[1-9]\d*"}, methods={"GET","PUT"})
      */
     public function update(Request $request, Category $category, CategoryRepository $repository, int $id)
     {
         $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        $category = $this->getCategory($id);
+        $category->getId();
         $form = $this->createForm(CategoryType::class, $category, ['method' => 'PUT']);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -130,10 +114,9 @@ class CategoryController extends AbstractController
         return $this->render('category/update.html.twig', [
             'id' => $id,
             'form' => $form->createView(),
-            'category' => $category
+            'category' => $category,
         ]);
     }
-
 
     /**
      * Delete action.
@@ -144,12 +127,20 @@ class CategoryController extends AbstractController
      *
      * @return RedirectResponse|\Symfony\Component\HttpFoundation\Response
      *
-     * @Route("/category/{id}", name="category_delete", requirements={"id": "[1-9]\d*"}, methods={"DELETE"})
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
      *
+     * @Route("/category/{id}", name="category_delete", requirements={"id": "[1-9]\d*"}, methods={"GET", "DELETE"})
      */
     public function delete(Request $request, Category $category, CategoryRepository $repository)
     {
         $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        if ($category->getPostings()->count()) {
+            $this->addFlash('danger', 'message.category_not_empty');
+
+            return $this->redirectToRoute('categories');
+        }
 
         $form = $this->createForm(CategoryType::class, $category, ['method' => 'DELETE']);
         $form->handleRequest($request);
@@ -173,31 +164,5 @@ class CategoryController extends AbstractController
                 'category' => $category,
             ]
         );
-
-    }
-
-    /**
-     * Repository getter.
-     *
-     * @return \Doctrine\Persistence\ObjectRepository
-     */
-    private function getRepository(): \Doctrine\Persistence\ObjectRepository
-    {
-        return $this->getDoctrine()->getRepository(Category::class);
-    }
-
-    /**
-     * Category getter.
-     *
-     * @param int $id Cateogry $id
-     *
-     * @return Category object
-     */
-    private function getCategory(int $id): object
-    {
-        $doctrineRepo = $this->getRepository();
-        $category = $doctrineRepo->find($id);
-
-        return $category;
     }
 }
